Updated IComponentManager interface and implemented behaviour Updated add/get components strategy Added Instantiatable.java interface Added ComponentType.java enum class Added joml library for calculations

+++ src/main/java/ecs/systems/processes/IUpdate.java
@@ -9,6 +9,6 @@ public interface IUpdate extends IProcess {
     /**
      * @param deltaTime is a time interval between present and previous frames
      */
-    default void update(double deltaTime) {
+    default void update(float deltaTime) {
     }
 }
